Rename NpgsqlUuid7ValueGenerator to NpgsqlSequentialGuidValueGenerator and make it public (#3255)
Follow-up to #3249
