+++ src/EFCore.PG/ValueGeneration/Internal/NpgsqlValueGeneratorSelector.cs
@@ -104,6 +104,6 @@ public class NpgsqlValueGeneratorSelector : RelationalValueGeneratorSelector
         => property.ClrType.UnwrapNullableType() == typeof(Guid)
             ? property.ValueGenerated == ValueGenerated.Never || property.GetDefaultValueSql() is not null
                 ? new TemporaryGuidValueGenerator()
-                : new NpgsqlUuid7ValueGenerator()
+                : new NpgsqlSequentialGuidValueGenerator()
             : base.FindForType(property, typeBase, clrType);
 }
