+++ test/EFCore.PG.Tests/NpgsqlValueGeneratorSelectorTest.cs
@@ -1,6 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.EntityFrameworkCore.ValueGeneration.Internal;
 using Npgsql.EntityFrameworkCore.PostgreSQL.TestUtilities;
+using Npgsql.EntityFrameworkCore.PostgreSQL.ValueGeneration;
 using Npgsql.EntityFrameworkCore.PostgreSQL.ValueGeneration.Internal;
 
 namespace Npgsql.EntityFrameworkCore.PostgreSQL;
@@ -21,7 +22,7 @@ public class NpgsqlValueGeneratorSelectorTest
         AssertGenerator<TemporaryByteValueGenerator>("NullableByte");
         AssertGenerator<TemporaryDecimalValueGenerator>("Decimal");
         AssertGenerator<StringValueGenerator>("String");
-        AssertGenerator<NpgsqlUuid7ValueGenerator>("Guid");
+        AssertGenerator<NpgsqlSequentialGuidValueGenerator>("Guid");
         AssertGenerator<BinaryValueGenerator>("Binary");
     }
 
@@ -128,7 +129,7 @@ public class NpgsqlValueGeneratorSelectorTest
         AssertGenerator<NpgsqlSequenceHiLoValueGenerator<long>>("NullableLong", setSequences: true);
         AssertGenerator<NpgsqlSequenceHiLoValueGenerator<short>>("NullableShort", setSequences: true);
         AssertGenerator<StringValueGenerator>("String", setSequences: true);
-        AssertGenerator<NpgsqlUuid7ValueGenerator>("Guid", setSequences: true);
+        AssertGenerator<NpgsqlSequentialGuidValueGenerator>("Guid", setSequences: true);
         AssertGenerator<BinaryValueGenerator>("Binary", setSequences: true);
     }
 
@@ -216,7 +217,7 @@ public class NpgsqlValueGeneratorSelectorTest
     {
         var dtoNow = DateTimeOffset.UtcNow;
         var net9Internal = Guid.CreateVersion7(dtoNow);
-        var custom = NpgsqlUuid7ValueGenerator.BorrowedFromNet9.CreateVersion7(dtoNow);
+        var custom = NpgsqlSequentialGuidValueGenerator.BorrowedFromNet9.CreateVersion7(dtoNow);
         var bytenet9 = net9Internal.ToByteArray().AsSpan(0, 6);
         var bytecustom = custom.ToByteArray().AsSpan(0, 6);
         Assert.Equal(bytenet9, bytecustom);
